analzer/requirements.txt
torch>=2.2.0
transformers>=4.36.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
nltk>=3.8.0
flask>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
Werkzeug>=2.0.0
pytesseract>=0.3.13
